move legend defaults to own module
